[DAD-856] feat: board와 요키 연결
