+++ src/pages/BoardInfoPage.tsx
@@ -1,9 +1,25 @@
 import { TrashableItems } from "@/components/templates/TrashableItems";
 import { useBoardAtom } from "@/hooks/useBoardAtom";
 import { useModal } from "@/hooks/useModal";
+import { YORKIE_API_KEY } from "@/secret";
 import { Box, Button, Typography } from "@mui/material";
 import { useLayoutEffect } from "react";
 import { useSearchParams } from "react-router-dom";
+import yorkie from "yorkie-js-sdk";
+
+async function connectYorkie() {
+    const client = new yorkie.Client('https://api.yorkie.dev', {
+        apiKey: YORKIE_API_KEY,
+    });
+    await client.activate();
+
+    const doc = new yorkie.Document('my-first-document');
+    await client.attach(doc);
+
+    return doc;
+}
+
+const doc = await connectYorkie();
 
 function BoardInfoPage() {
     const [searchParams, setSearchParams] = useSearchParams();
@@ -13,12 +29,13 @@ function BoardInfoPage() {
     }
 
     const { board, setBoard } = useBoardAtom();
+    const boardPageId = getBoardPageId();
 
     useLayoutEffect(() => {
         setBoard({
             ...board,
-            title: '보드 1',
-            boardId: getBoardPageId(),
+            title: doc.getRoot()[boardPageId].title,
+            boardId: boardPageId,
         });
     }, [])
 
@@ -47,7 +64,7 @@ function BoardInfoPage() {
                     justifyContent: 'space-between',
                 }}
             >
-                <TrashableItems confirmDrop={false} />
+                <TrashableItems confirmDrop={false} doc={doc} boardId={boardPageId} />
                 <Box>
                     <Button
                         onClick={() => openModal('scrapCreateOnBoard')}
